+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to
 ## [Unreleased]
 
 - Documenting support for python 3.13. (#86)
+- Only show failing rules per default in `HumanReadableFormatter`. Also added
+  `--show` parameter in the CLI to change this behavior. (#77)
 
 ## [0.8.0] - 2024-11-12